vault-agent/pyproject.toml
[project]
name = "vault-agent"
version = "0.1.0"
description = "Vault PKI Query Agent - Natural language interface for HashiCorp Vault PKI operations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "strands-agents[openai]>=1.13.0",
    "strands-agents-tools>=0.2.12",
    "streamlit>=1.28.0",
    "httpx>=0.25.0",
    "pydantic>=2.4.0",
    "python-dotenv>=1.0.0",
    "openai>=1.109.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]
